changed quantity data type to int

+++ data_pipeline.py
@@ -96,7 +96,7 @@ cursor.execute('''
     CREATE TABLE IF NOT EXISTS sales (
         transaction_id TEXT PRIMARY KEY,
         product_id TEXT,
-        quantity REAL,
+        quantity INTEGER,
         price REAL,
         timestamp TEXT
     )
